Day9/video-avatars-agent/agents/video_avatar_agent/requirements.txt
google-adk==1.17.*
google-cloud-aiplatform==1.125.*
google-cloud-storage
httpx==0.28.*
uvicorn==0.38.*
